requirements.txt
# Core dependencies - tested with Python 3.14
fastapi==0.121.3
uvicorn[standard]==0.38.0
openai==2.8.1
anthropic==0.74.1
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.4
pydantic-settings==2.12.0
boto3==1.35.76  # For AWS Secrets Manager integration
requests-aws4auth==1.3.1  # For IAM-authenticated API Gateway requests

# LangChain and LangGraph for agentic workflows
langchain>=1.1.0
langchain-openai>=1.1.0
langchain-anthropic>=1.2.0
langchain-core>=1.1.0
langgraph>=1.0.4

# Testing
pytest>=8.0.0
